24.9.27庄登录注册

+++ src/main/java/com/team8504/csgo2web/entity/News.java
@@ -7,8 +7,8 @@ import lombok.ToString;
 @ToString
 public class News {
 
-  private long nId;
-  private long cId;
+  private Integer nId;
+  private Integer cId;
   private String nTitle;
   private String nDesc;
   private String nThumb;
